requirements.txt
python-telegram-bot==20.7
python-dotenv>=1.0.0
gspread>=6.0.0          # Google Sheets API (автоматически ставит oauth2client)
oauth2client>=4.1.3
loguru>=0.7.0
aiofiles>=22.1.0
requests>=2.28.0